Fixed issue #2099.
