+++ javaparser-core-testing/src/test/java/com/github/javaparser/ast/stmt/BlockStmtTest.java
@@ -1,6 +1,8 @@
 package com.github.javaparser.ast.stmt;
 
 import com.github.javaparser.JavaParser;
+import com.github.javaparser.ParserConfiguration;
+import com.github.javaparser.StaticJavaParser;
 import com.github.javaparser.ast.expr.Expression;
 import com.github.javaparser.ast.expr.MethodCallExpr;
 import com.github.javaparser.ast.expr.NameExpr;
@@ -23,33 +25,4 @@ class BlockStmtTest {
         // This fails when the issue exists:
         String s = blockStmt.toString();
     }
-
-    @Test
-    void issue2099AddingAddingStatementAfterTraillingComment() {
-        JavaParser javaParser = new JavaParser();
-
-        Statement statement = javaParser.parseStatement(
-            "if(value != null) {" + EOL +
-            "    value.value();" + EOL +
-            "}").getResult().get();
-
-        BlockStmt blockStmt = LexicalPreservingPrinter.setup(
-                javaParser.parseBlock("{" + EOL +
-                "    value(); // Test" + EOL +
-                "}").getResult().get());
-
-        blockStmt.addStatement(statement);
-        String s = LexicalPreservingPrinter.print(blockStmt);
-        String expected = "{\n" +
-                "    value(); // Test\n" +
-                "    if(value != null) {\n" +
-                "        value.value();\n" +
-                "    }\n" +
-                "}";
-        System.out.println("EXPECTED:\n" + expected);
-        System.out.println("RESULT:\n" + s);
-
-        assertEqualsNoEol(expected, blockStmt.toString());
-    }
-
 }
